pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "procureneg-gym"
version = "0.1.0"
description = "Procurement Negotiation RL Environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn",
  "pydantic",
  "pyyaml",
  "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"
